+++ internal/agdcslog/syslog_linux.go
@@ -1,4 +1,4 @@
-//go:build !windows
+//go:build linux
 
 package agdcslog
 
@@ -6,13 +6,13 @@ import (
 	"log/syslog"
 )
 
-// systemLogger is a wrapper around [eventlog.Log].
+// systemLogger is the implementation of the [SystemLogger] interface for Linux.
 type systemLogger struct {
 	writer *syslog.Writer
 }
 
-// newSystemLogger returns a unix specific system logger.
-func newSystemLogger(tag string) (l Logger, err error) {
+// newSystemLogger returns a Linux-specific system logger.
+func newSystemLogger(tag string) (l SystemLogger, err error) {
 	const priority = syslog.LOG_NOTICE | syslog.LOG_USER
 
 	w, err := syslog.New(priority, tag)
@@ -27,29 +27,29 @@ func newSystemLogger(tag string) (l Logger, err error) {
 }
 
 // type check
-var _ Logger = (*systemLogger)(nil)
+var _ SystemLogger = (*systemLogger)(nil)
 
-// Debug implements [Logger] interface for *systemLogger.
+// Debug implements the [SystemLogger] interface for *systemLogger.
 func (l *systemLogger) Debug(msg string) (err error) {
 	return l.writer.Debug(msg)
 }
 
-// Info implements [Logger] interface for *systemLogger.
+// Info implements the [SystemLogger] interface for *systemLogger.
 func (l *systemLogger) Info(msg string) (err error) {
 	return l.writer.Info(msg)
 }
 
-// Warning implements [Logger] interface for *systemLogger.
+// Warning implements the [SystemLogger] interface for *systemLogger.
 func (l *systemLogger) Warning(msg string) (err error) {
 	return l.writer.Warning(msg)
 }
 
-// Error implements [Logger] interface for *systemLogger.
+// Error implements the [SystemLogger] interface for *systemLogger.
 func (l *systemLogger) Error(msg string) (err error) {
 	return l.writer.Err(msg)
 }
 
-// Close implements [Logger] interface for *systemLogger.
+// Close implements the [SystemLogger] interface for *systemLogger.
 func (l *systemLogger) Close() (err error) {
 	return l.writer.Close()
 }